+++ querybuilder/tests/select_tests.py
@@ -105,7 +105,11 @@ class SelectTest(QueryTestCase):
             }]
         )
         query_str = query.get_sql()
-        expected_query = 'SELECT test_table.field_one AS field_alias_one, test_table.field_two AS field_alias_two FROM test_table'
+        expected_query = (
+            'SELECT test_table.field_one AS field_alias_one, '
+            'test_table.field_two AS field_alias_two '
+            'FROM test_table'
+        )
         self.assertEqual(query_str, expected_query, get_comparison_str(query_str, expected_query))
 
     def test_select_fields_alias_from_string_alias(self):
@@ -120,7 +124,12 @@ class SelectTest(QueryTestCase):
             }]
         )
         query_str = query.get_sql()
-        expected_query = 'SELECT table_alias.field_one AS field_alias_one, table_alias.field_two AS field_alias_two FROM test_table AS table_alias'
+        expected_query = (
+            'SELECT table_alias.field_one AS field_alias_one, '
+            'table_alias.field_two AS field_alias_two '
+            'FROM test_table '
+            'AS table_alias'
+        )
         self.assertEqual(query_str, expected_query, get_comparison_str(query_str, expected_query))
 
     def test_select_fields_alias_from_model(self):
@@ -133,7 +142,11 @@ class SelectTest(QueryTestCase):
             }]
         )
         query_str = query.get_sql()
-        expected_query = 'SELECT tests_account.field_one AS field_alias_one, tests_account.field_two AS field_alias_two FROM tests_account'
+        expected_query = (
+            'SELECT tests_account.field_one AS field_alias_one, '
+            'tests_account.field_two AS field_alias_two '
+            'FROM tests_account'
+        )
         self.assertEqual(query_str, expected_query, get_comparison_str(query_str, expected_query))
 
     def test_select_fields_alias_from_model_alias(self):
@@ -148,7 +161,12 @@ class SelectTest(QueryTestCase):
             }]
         )
         query_str = query.get_sql()
-        expected_query = 'SELECT table_alias.field_one AS field_alias_one, table_alias.field_two AS field_alias_two FROM tests_account AS table_alias'
+        expected_query = (
+            'SELECT table_alias.field_one AS field_alias_one, '
+            'table_alias.field_two AS field_alias_two '
+            'FROM tests_account '
+            'AS table_alias'
+        )
         self.assertEqual(query_str, expected_query, get_comparison_str(query_str, expected_query))
 
     def test_select_fields_two_tables(self):
@@ -166,7 +184,13 @@ class SelectTest(QueryTestCase):
             ]
         )
         query_str = query.get_sql()
-        expected_query = 'SELECT tests_account.field_one, tests_account.field_two, second_table.field_three, second_table.field_four FROM tests_account, second_table'
+        expected_query = (
+            'SELECT tests_account.field_one, '
+            'tests_account.field_two, '
+            'second_table.field_three, '
+            'second_table.field_four '
+            'FROM tests_account, second_table'
+        )
         self.assertEqual(query_str, expected_query, get_comparison_str(query_str, expected_query))
 
     def test_select_fields_two_tables_alias(self):
@@ -190,7 +214,14 @@ class SelectTest(QueryTestCase):
             }]
         )
         query_str = query.get_sql()
-        expected_query = 'SELECT table_one.field_one AS f1, table_one.field_two AS f2, table_two.field_three AS f3, table_two.field_four AS f4 FROM tests_account AS table_one, second_table AS table_two'
+        expected_query = (
+            'SELECT table_one.field_one AS f1, '
+            'table_one.field_two AS f2, '
+            'table_two.field_three AS f3, '
+            'table_two.field_four AS f4 '
+            'FROM tests_account AS table_one, '
+            'second_table AS table_two'
+        )
         self.assertEqual(query_str, expected_query, get_comparison_str(query_str, expected_query))
 
     def test_select_fields_same_two_tables(self):
@@ -208,7 +239,13 @@ class SelectTest(QueryTestCase):
             ]
         )
         query_str = query.get_sql()
-        expected_query = 'SELECT tests_account.field_one, tests_account.field_two, T1.field_three, T1.field_four FROM tests_account, tests_account AS T1'
+        expected_query = (
+            'SELECT tests_account.field_one, '
+            'tests_account.field_two, '
+            'T1.field_three, '
+            'T1.field_four '
+            'FROM tests_account, tests_account AS T1'
+        )
         self.assertEqual(query_str, expected_query, '\n{0}\n!=\n{1}'.format(query_str, expected_query))
 
 
@@ -257,7 +294,11 @@ class InnerQueryTest(QueryTestCase):
         )
 
         query_str = query.get_sql()
-        expected_query = 'WITH T0 AS (SELECT tests_account.* FROM tests_account WHERE (id > %(T0A0)s AND id < %(T0A1)s)) SELECT T0.* FROM T0'
+        expected_query = (
+            'WITH T0 AS '
+            '(SELECT tests_account.* FROM tests_account WHERE (id > %(T0A0)s AND id < %(T0A1)s)) '
+            'SELECT T0.* FROM T0'
+        )
         self.assertEqual(query_str, expected_query, get_comparison_str(query_str, expected_query))
 
     def test_inner_outer_args(self):
@@ -273,7 +314,11 @@ class InnerQueryTest(QueryTestCase):
         )
 
         query_str = query.get_sql()
-        expected_query = 'WITH T0 AS (SELECT tests_account.* FROM tests_account WHERE (id > %(T0A0)s AND id < %(T0A1)s)) SELECT T0.* FROM T0 WHERE ((NOT(id = %(A0)s)))'
+        expected_query = (
+            'WITH T0 AS '
+            '(SELECT tests_account.* FROM tests_account WHERE (id > %(T0A0)s AND id < %(T0A1)s)) '
+            'SELECT T0.* FROM T0 WHERE ((NOT(id = %(A0)s)))'
+        )
         self.assertEqual(query_str, expected_query, get_comparison_str(query_str, expected_query))
 
     def test_inner_outer_args_many(self):
@@ -298,7 +343,17 @@ class InnerQueryTest(QueryTestCase):
         )
 
         query_str = query.get_sql()
-        expected_query = 'WITH T1 AS (SELECT tests_account.* FROM tests_account WHERE (id > %(T1A0)s AND id < %(T1A1)s)), T0 AS (SELECT tests_account.* FROM tests_account WHERE (id > %(T0A0)s AND id < %(T0A1)s)) SELECT T0.*, T1.* FROM T0, T1 WHERE ((NOT(id = %(A0)s)))'
+        expected_query = (
+            'WITH T1 AS '
+            '(SELECT tests_account.* FROM tests_account WHERE (id > %(T1A0)s AND id < %(T1A1)s)), '
+            'T0 AS ('
+            'SELECT tests_account.* '
+            'FROM tests_account '
+            'WHERE (id > %(T0A0)s AND id < %(T0A1)s)) '
+            'SELECT T0.*, T1.* '
+            'FROM T0, T1 '
+            'WHERE ((NOT(id = %(A0)s)))'
+        )
         self.assertEqual(query_str, expected_query, get_comparison_str(query_str, expected_query))
 
     def test_three_levels(self):
@@ -336,7 +391,13 @@ class InnerQueryTest(QueryTestCase):
             ~Q(id=0)
         )
         query_str = query.get_sql()
-        expected_query = 'WITH T1T1 AS (SELECT tests_account.* FROM tests_account WHERE (id > %(T1T1A0)s AND id < %(T1T1A1)s)), T1T0 AS (SELECT tests_account.* FROM tests_account WHERE (id > %(T1T0A0)s AND id < %(T1T0A1)s)), T1 AS (SELECT T1T0.*, T1T1.* FROM T1T0, T1T1 WHERE (id > %(T1A0)s AND id < %(T1A1)s)), T0 AS (SELECT tests_account.* FROM tests_account WHERE (id > %(T0A0)s AND id < %(T0A1)s)) SELECT T0.*, T1.* FROM T0, T1 WHERE ((NOT(id = %(A0)s)))'
+        expected_query = (
+            'WITH T1T1 AS (SELECT tests_account.* FROM tests_account WHERE (id > %(T1T1A0)s AND id < %(T1T1A1)s)), '
+            'T1T0 AS (SELECT tests_account.* FROM tests_account WHERE (id > %(T1T0A0)s AND id < %(T1T0A1)s)), '
+            'T1 AS (SELECT T1T0.*, T1T1.* FROM T1T0, T1T1 WHERE (id > %(T1A0)s AND id < %(T1A1)s)), '
+            'T0 AS (SELECT tests_account.* FROM tests_account WHERE (id > %(T0A0)s AND id < %(T0A1)s)) '
+            'SELECT T0.*, T1.* FROM T0, T1 WHERE ((NOT(id = %(A0)s)))'
+        )
         self.assertEqual(query_str, expected_query, get_comparison_str(query_str, expected_query))