+++ run_tests.py
@@ -1,7 +1,6 @@
 """
 Provides the ability to run test on a standalone Django app.
 """
-import os
 import sys
 from optparse import OptionParser