+++ docs/release_notes.rst
@@ -1,6 +1,12 @@
 Release Notes
 =============
 
+v0.5.8
+------
+
+* Added connection parameter to query for multiple database support
+* flake8 cleanup
+
 v0.5.7
 ------